Add reference resolve support to color commands
