+++ gradle.properties
@@ -1,4 +1,4 @@
-pluginVersion = 1.1.1-alpha.1
+pluginVersion = 1.1.1-alpha.2
 
 #     Info about build ranges: https://www.jetbrains.org/intellij/sdk/docs/basics/getting_started/build_number_ranges.html
 #    Note that an xyz branch corresponds to version 20xy.z and a since build of xyz.*
